requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
pymysql
python-dotenv
python-multipart
requests
email-validator
